+++ crates/vertigo/src/tests/mod.rs
@@ -1,8 +1,8 @@
 // Tests that can't be placed directly alongside the code,
 // for example tests of macros.
 
-mod api_access;
 mod autojsjson;
 mod bind;
 mod css;
 mod dom;
+mod js_macro;